Move Bookmark model to the domain package
Same as LikedTweet: entity models live in domain, the bookmark repo,
handler and tests now reference domain.Bookmark.
